+++ src/run.py
@@ -18,10 +18,14 @@
 from __future__ import annotations
 
 import argparse
+import json
 from dataclasses import dataclass
 from pathlib import Path
 
 import jinja2
+
+DEFAULTS_DIR = Path(__file__).absolute().parent.parent / "defaults"
+TEMPLATES_DIR = Path(__file__).absolute().parent.parent / "templates"
 
 
 @dataclass
@@ -32,6 +36,12 @@ class TemplateContainer:
     output_dir: Path
     active: bool
     arguments: dict[str, str]
+    template_name: str | None = None
+
+    def __post_init__(self) -> None:
+        """Set the template name if not provided."""
+        if self.template_name is None:
+            self.template_name = self.file_name
 
 
 def main(
@@ -41,10 +51,16 @@ def main(
     name: str,
     synchronize_issue_templates: bool,
     synchronize_pull_request_template: bool,
+    synchronize_release_drafter_template: bool,
     synchronize_security_policy: bool,
     synchronize_support_resources: bool,
+    release_drafter_categories: str,
 ) -> None:
     """Render all templates."""
+    if not release_drafter_categories:
+        release_drafter_categories = Path(DEFAULTS_DIR / "release_drafter_categories.json").read_text(encoding="utf-8")
+    release_drafter_categories_dict = json.loads(release_drafter_categories)
+
     template_containers = [
         TemplateContainer(
             file_name="bug-report.yml",
@@ -65,6 +81,13 @@ def main(
             arguments={},
         ),
         TemplateContainer(
+            file_name="release-drafter.yml",
+            template_name="release-drafter.yml.in",
+            output_dir=Path(".github"),
+            active=synchronize_release_drafter_template,
+            arguments={"name": name, "release_drafter_categories": release_drafter_categories_dict},
+        ),
+        TemplateContainer(
             file_name="SECURITY.md",
             output_dir=Path(".github"),
             active=synchronize_security_policy,
@@ -79,7 +102,7 @@ def main(
     ]
 
     environment = jinja2.Environment(
-        loader=jinja2.FileSystemLoader(Path(__file__).absolute().parent.parent / "templates"),
+        loader=jinja2.FileSystemLoader(TEMPLATES_DIR),
         autoescape=jinja2.select_autoescape(["html", "xml", "htm"]),
     )
 
@@ -90,9 +113,16 @@ def main(
 def _render_template(environment: jinja2.Environment, template_container: TemplateContainer) -> None:
     """Render a template."""
     if template_container.active:
-        template = environment.get_template(template_container.file_name)
+        # Create output directory if it does not exist
+        output_dir = template_container.output_dir
+        output_dir.mkdir(parents=True, exist_ok=True)
+
+        # Render the template
+        template = environment.get_template(template_container.template_name)
         output = template.render(**template_container.arguments)
-        output_path = template_container.output_dir / template_container.file_name
+
+        # Write the rendered rendered to the file
+        output_path = output_dir / template_container.file_name
         output_path.write_text(output + "\n", encoding="utf-8")
 
 
@@ -138,6 +168,12 @@ if __name__ == "__main__":
         help="Whether to synchronize the pull-request template",
     )
     parser.add_argument(
+        "--synchronize_release_drafter_template",
+        default=True,
+        type=_convert_to_bool,
+        help="Whether to synchronize the Release Drafter template",
+    )
+    parser.add_argument(
         "--synchronize_security_policy",
         default=True,
         type=_convert_to_bool,
@@ -149,6 +185,12 @@ if __name__ == "__main__":
         type=_convert_to_bool,
         help="Whether to synchronize the support resources",
     )
+    parser.add_argument(
+        "--release_drafter_categories",
+        type=str,
+        default="",
+        help="Release Drafter categories as a JSON string. If empty, a reasonable default will be used",
+    )
     args = parser.parse_args()
 
     main(
@@ -157,6 +199,8 @@ if __name__ == "__main__":
         name=args.name,
         synchronize_issue_templates=args.synchronize_issue_templates,
         synchronize_pull_request_template=args.synchronize_pull_request_template,
+        synchronize_release_drafter_template=args.synchronize_release_drafter_template,
         synchronize_security_policy=args.synchronize_security_policy,
         synchronize_support_resources=args.synchronize_support_resources,
+        release_drafter_categories=args.release_drafter_categories,
     )